pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "catalogic"
version = "0.1.0"
description = "Умный мультиплатформенный каталогизатор файлов"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "PolyForm Noncommercial 1.0.0"}
dependencies = [
    "python-magic>=0.4.27",
    "Pillow>=10",
    "fastapi>=0.115",
    "uvicorn>=0.30",
]

[project.optional-dependencies]
dev = ["pytest>=7"]

[project.scripts]
catalogic = "catalogic.cli:main"
catalogic-server = "catalogic.server:main"
catalogic-scanner = "catalogic.agent.runner:main"
catalogic-frontend = "catalogic.frontend.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
catalogic = ["web/*.html", "web/static/*.css", "web/static/*.js"]
